requirements.txt
Flask
opencv-python
numpy
pillow
werkzeug
